engine/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.9
pyyaml>=6.0

# ML Classifier
scikit-learn>=1.3.0
numpy>=1.24.0

# RAG + Agent
openai>=1.0.0
faiss-cpu>=1.7.0

# Optional: better embeddings (uncomment for better RAG)
# sentence-transformers>=2.2.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
